Fix 3 UBSan signed overflows in trade route scoring and city specialization
ScoreInternationalTR (CvTradeClasses.cpp):
  Widen int iScore to long long. The function accumulates iScore through
  ~15 conditional multiplicative adjustments (×25 per city in a loop,
  ×10 feitoria, ×EventTourismCS, ×YieldPerFranchise, etc.) that compound
  well past INT_MAX in mid-to-late game. Saturate at INT_MAX when
  assigning to ret.m_iScore (int, transient, not serialized).

  Fixes two UBSan findings:
    270000000 * 8 at line 6391 (T+0.016s, pre-Turn 0)
    928000000 * 5 at line 6426 (T+1858s, Turn 97)

  Design debt: the *=25 inside a city loop (line 6385) multiplies per
  qualifying city, not once — probable gameplay logic bug causing
  exponential compounding and degenerate ranking post-saturation.
  Documented in ubsan-notes Section 12.3.

AssignSpecializations (CvCitySpecializationAI.cpp):
  Widen int iCityValue and int iBestValue to long long. This is
  downstream of the prior AdjustValueBasedOnBuildings fix (2bf1bb4f7)
  which saturates at INT_MAX — when summing per-yield values for
  GENERAL_ECONOMIC specialization, adding INT_MAX (Production) to the
  accumulator overflows.

  Fixes one UBSan finding:
    92787 + 2147483647 at line 735 (T+1182s, Turn 72)

  Design debt: AdjustValueBasedOnBuildings chained percentage pattern
  routinely exceeds INT_MAX for late-game production cities. Runtime
  logs confirm 7+ cities at INT_MAX production (turns 203+). Ranking
  degeneracy documented in ubsan-notes Section 12.4.

Gameplay behavior change: previously overflow wrapped to arbitrary
(potentially negative) values, corrupting rankings non-deterministically.
Post-fix, values saturate deterministically at INT_MAX. This is an
improvement but causes ranking degeneracy when multiple candidates
saturate. No save, ABI, or network protocol impact.

+++ CvGameCoreDLL_Expansion2/CvCitySpecializationAI.cpp
@@ -716,7 +716,7 @@ void CvCitySpecializationAI::AssignSpecializations()
 		bool bCoastal = pkCitySpecializationEntry->IsMustBeCoastal();
 
 		// Pick best existing city based on a better computation of existing city's value for a yield type
-		int iBestValue = 1;
+		long long iBestValue = 1;
 		int iBestCityID = -1;
 		for (map<int, vector<int>>::iterator itCity = citiesWithoutSpecialization.begin(); itCity != citiesWithoutSpecialization.end(); ++itCity)
 		{
@@ -728,7 +728,7 @@ void CvCitySpecializationAI::AssignSpecializations()
 			if (pCity->getHappinessDelta()<0 && eYield == YIELD_FOOD)
 				continue;
 
-			int iCityValue = 0;
+			long long iCityValue = 0;
 			if (eYield == NO_YIELD)
 			{
 				for (int iI = 0; iI < YIELD_TOURISM; iI++)
